added support for sanity check
Sanity check mode validates whether the YAML files with known issues are
consistent.

+++ knownissues/sync-known-issues.py
@@ -34,7 +34,12 @@ class SquadConnection(object):
         }
 
     def get_prepared_request(self, endpoint, method):
-        URL = urlunsplit((self.url_scheme, self.base_url, "api/%s" % endpoint, None, None))
+        URL = urlunsplit(
+            (self.url_scheme,
+             self.base_url,
+             "api/%s" % endpoint,
+             None,
+             None))
         req = requests.Request(method, URL, headers=self.headers)
         return req.prepare()
 
@@ -49,7 +54,12 @@ class SquadConnection(object):
             sys.exit(1)
 
     def download_list(self, endpoint, params=None):
-        URL = urlunsplit((self.url_scheme, self.base_url, "api/%s" % endpoint, None, None))
+        URL = urlunsplit(
+            (self.url_scheme,
+             self.base_url,
+             "api/%s" % endpoint,
+             None,
+             None))
         logger.debug(URL)
         response = requests.get(URL, params=params, headers=self.headers)
         result_list = []
@@ -90,7 +100,12 @@ class SquadConnection(object):
 
     def put_object(self, endpoint, config):
         object_id = config.get('id')
-        URL = urlunsplit((self.url_scheme, self.base_url, "api/%s/%s/" % (endpoint, object_id), None, None))
+        URL = urlunsplit(
+            (self.url_scheme,
+             self.base_url,
+             "api/%s/%s/" % (endpoint, object_id),
+             None,
+             None))
         logger.debug(URL)
         logger.debug(config)
         response = requests.put(URL, data=config, headers=self.headers)
@@ -111,7 +126,7 @@ class SquadKnownIssueException(Exception):
 
 
 class SquadKnownIssue(object):
-    def __init__(self, config, squad_project):
+    def __init__(self, config, squad_project, strict=False):
         self.test_name = config.get('test_name')
         if self.test_name is None:
             raise SquadKnownIssueException("TestName not defined")
@@ -126,13 +141,25 @@ class SquadKnownIssue(object):
                 # ignore projects that are not defined
                 # in the SquadProject.projects
                 self.projects.remove(project)
+                if strict:
+                    raise SquadKnownIssueException("Project not defined: %s" % project)
         self.environments = set()
         for item in config.get('environments'):
-            if 'slug' in item.keys() and item['slug'] in squad_project.environment_slugs:
-                self.environments.add(item['slug'])
-            if 'architecture' in item.keys() and item['architecture'] in squad_project.environment_architectures.keys():
-                for env in squad_project.environment_architectures[item['architecture']]:
-                    self.environments.add(env)
+            if 'slug' in item.keys():
+                if item['slug'] in squad_project.environment_slugs:
+                    self.environments.add(item['slug'])
+                else:
+                    if strict:
+                        raise SquadKnownIssueException(
+                            "Incorrect environment: %s" % item['slug'])
+            if 'architecture' in item.keys():
+                if item['architecture'] in squad_project.environment_architectures.keys():
+                    for env in squad_project.environment_architectures[item['architecture']]:
+                        self.environments.add(env)
+                else:
+                    if strict:
+                        raise SquadKnownIssueException(
+                            "Unknown architecture: %s" % item['architecture'])
 
 
 class SquadProjectException(Exception):
@@ -140,7 +167,7 @@ class SquadProjectException(Exception):
 
 
 class SquadProject(object):
-    def __init__(self, config, passwords_file):
+    def __init__(self, config, passwords_file, sanity_check=False):
         self.name = config.get('name')
         self.url = config.get('url')
         if self.url is None:
@@ -158,7 +185,7 @@ class SquadProject(object):
                 self.environment_architectures[arch_name].add(item.get('slug'))
             else:
                 self.environment_architectures.update({arch_name: set([item.get('slug')])})
-        self.known_issues = [SquadKnownIssue(conf, self) for conf in config.get('known_issues')]
+        self.known_issues = [SquadKnownIssue(conf, self, sanity_check) for conf in config.get('known_issues')]
 
 
 def main():
@@ -180,6 +207,12 @@ def main():
                         default=False,
                         help="Dry run",
                         dest="dry_run")
+    parser.add_argument("-s",
+                        "--sanity-check",
+                        action="store_true",
+                        default=False,
+                        help="Sanity check. Implies dry run.",
+                        dest="sanity_check")
     parser.add_argument("-v",
                         "--debug",
                         action="store_true",
@@ -200,53 +233,69 @@ def main():
                 logger.error(exc)
                 sys.exit(1)
 
-    pprint.pprint(config_data)
     for project_name, project in config_data.items():
-        s = SquadProject(project, args.passwords_file)
-        print(s.environment_architectures)
+        s = SquadProject(project, args.passwords_file, args.sanity_check)
+        if args.sanity_check:
+            # validate if projects defined in the instance exist
+            for squad_project in s.projects:
+                squad_project_group, squad_project_name = squad_project.split("/", 1)
+                api_project = s.connection.filter_object(
+                    'projects',
+                    {'group__slug': squad_project_group, 'slug': squad_project_name})
+                if api_project is None:
+                    raise SquadProjectException(
+                        "Project %s doesn't exist in the instance %s" % (squad_project, s.url))
+
         for known_issue in s.known_issues:
             # create/uptade the issues in remote instance
-            print(known_issue.title)
-            print(known_issue.test_name)
-            print(known_issue.environments)
             # for each project defined in the known_issue
             # get the environment IDs based on the environment name
-            api_known_issue = s.connection.filter_object('knownissues', {'title': known_issue.title, 'test_name': known_issue.test_name})
+            api_known_issue = s.connection.filter_object(
+                'knownissues',
+                {'title': known_issue.title, 'test_name': known_issue.test_name})
             affected_environments = []
             for known_issue_project in known_issue.projects:
                 group_name, project_name = known_issue_project.split('/', 1)
-                api_project = s.connection.filter_object('projects', {'group__slug': group_name, 'slug': project_name})
+                api_project = s.connection.filter_object(
+                    'projects',
+                    {'group__slug': group_name, 'slug': project_name})
                 if api_project is None:
                     continue
-                api_environments = s.connection.download_list('environments', {'project': api_project['id']})
+                api_environments = s.connection.download_list(
+                    'environments',
+                    {'project': api_project['id']})
                 for api_env in api_environments:
                     if api_env['slug'] in known_issue.environments:
-                        logger.debug("Adding env: %s to known issue: %s" % (api_env['slug'], known_issue.title))
+                        logger.debug(
+                            "Adding env: %s to known issue: %s" % (
+                                api_env['slug'],
+                                known_issue.title))
                         affected_environments.append(api_env)
 
-            known_issue_api_object = {
-                'title': known_issue.title,
-                'test_name': known_issue.test_name,
-                'url': known_issue.url,
-                'notes': known_issue.notes,
-                'active': known_issue.active,
-                'intermittent': known_issue.intermittent,
-                'environment': [item['url'] for item in affected_environments]
-            }
+            if not args.dry_run:
+                known_issue_api_object = {
+                    'title': known_issue.title,
+                    'test_name': known_issue.test_name,
+                    'url': known_issue.url,
+                    'notes': known_issue.notes,
+                    'active': known_issue.active,
+                    'intermittent': known_issue.intermittent,
+                    'environment': [item['url'] for item in affected_environments]
+                }
 
-            if api_known_issue is None:
-                # create new KnownIssue
-                s.connection.post_object(
-                    'knownissues',
-                    known_issue_api_object
-                )
-            else:
-                # update existing KnownIssue
-                api_known_issue.update(known_issue_api_object)
-                s.connection.put_object(
-                    'knownissues',
-                    api_known_issue
-                )
+                if api_known_issue is None:
+                    # create new KnownIssue
+                    s.connection.post_object(
+                        'knownissues',
+                        known_issue_api_object
+                    )
+                else:
+                    # update existing KnownIssue
+                    api_known_issue.update(known_issue_api_object)
+                    s.connection.put_object(
+                        'knownissues',
+                        api_known_issue
+                    )
 
 
 if __name__ == '__main__':
